Refactored directories management, now uses seperates activities/models
