+++ app/src/main/java/org/wit/hillfortexplorer/activities/HillfortActivity.kt
@@ -1,4 +1,4 @@
-package org.wit.hillfortexplorer
+package org.wit.hillfortexplorer.activities
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -6,6 +6,7 @@ import kotlinx.android.synthetic.main.activity_hillfort.*
 import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.info
 import org.jetbrains.anko.toast
+import org.wit.hillfortexplorer.R
 
 class HillfortActivity : AppCompatActivity(), AnkoLogger {
