+++ app/src/main/java/org/wit/hillfortexplorer/models/HillfortModel.kt
@@ -1,0 +1,3 @@
+package org.wit.hillfortexplorer.models
+
+data class HillfortModel(var title: String = "")
