+++ app/src/main/res/layout/activity_hillfort.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HillfortActivity">
+    tools:context=".activities.HillfortActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
